requirements.txt
azure-identity>=1.15.0
python-dotenv>=1.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-monitor>=6.0.0
azure-monitor-query>=1.2.0
